+++ ov-budget/public/index.php
@@ -62,6 +62,7 @@ $routes = [
     'event_file'         => 'event_file',
     'event_print'        => 'event_print',
     'event_export'       => 'event_export',
+    'event_invites'      => 'event_invites',
 
     // Fahrzeuge
     'vehicles'           => 'vehicles',